requirements.txt
opencv-python>=4.8.0
numpy>=1.24.0
easyocr>=1.7.1
google-generativeai>=0.3.0
python-dotenv>=1.0.0
PyMuPDF>=1.23.0
Pillow==10.0.0
python-dateutil==2.8.2
PyPDF2==3.0.1 